Use *_PACK_OFFSET instead of pack_len for strncpy
When pre-tagging the buffer to store the packed representation of a
resource object, the code was using pack_len as the buffer size.
pack_len may not always be larger than the size of the tag offset, so
the appropriate *_PACK_OFFSET constant should be used instead, since it
is defined as the length of the pack tag.