Abstract Syntax Tree Comparison
The new ast_compare function compares two abstract syntax trees, node
for node, and returns 0 if they are both identical in content.  This is
intended to be used for policy versioning, but the necessary support
mechanisms in the policy module are not built yet.  Expect a forthcoming
commit to address this lack.