Resource Overrides
When a resource is defined later in a policy (or via an extend or
enforce), its attributes override and add to the previously definied
resource, in a form of inheritance.  This obviates the need for the
res_*_merge functions and any supporting macros (which have been
removed).  Also, the *_prio members are gone.