fact_read can be called in sequence
The second parameter to fact_read(FILE*, struct hash*) allows fact_read
to be called multiple times in sequence, to build up a single hash of
values.  This was implemented for the client utility, which will need to
execute a series of scripts and parse each ones output in turn.