Update tests for res_file pack changes
res_file_pack now packs the remote SHA1 instead of the remote
path, since the new worker thread approach allows the server to
search through the policy based on remote checksum.  The client
gets the remote checksum, and if it does not equal the local
checksum, the client sends a GET_FILE PDU, with the remote SHA1
hex value, to get the contents of the file in a series of
FILE_DATA PDUs.

This commit fixes the tests so that they do not expect the
remote path in the packed or unpacked output.