Allow individual unit test suites to be run
test/run now accepts a single argument, the name of a test
suite to run.  This argument is optional; if not specified, all
test suites will be run.  The Makefile has been updated to include
a new target, 'unit' that runs a single test suite named in the
TEST environment variable.