New 'U' Pendulum register for userdata
This makes it easier to keep handles and resources for things like the
passwd database, without cluttering up the main pn_machine definition
with lots of registers.

This also removes our reliance on %A to store a wild pointer to the last
found user, which cleans up the code by reducing register pressure.