Pendulum Heap support
The pn_machine VM now has a "heap" for tracking dynamically allocated
blocks of memory, and properly garbage-collecting them when there are no
references to them in the %A-%F, %S1, %S2, %T1 or %T2 registers.

This should allow us to safely compare the output of multiple EXEC.RUN1
calls, for use in package version comparisons, without leaking memory.