More Robust cogd/clockd testing
New test-only binaries, TEST_cogd and TEST_clockd, have some compiled
in features that we want in testing, but not in the real world,
including:

- A 60s failsafe that will terminate the process
- Single-run clockd (exit after first ERROR or BYE PDU sent)
- No root UID/EUID checking in cogd

With this new, better mechanism, previous mechanisms are deprected.
Notably, the COGD_SKIP_ROOT environment variable is no longer supported,
and the super.secret.dev.mode clockd configuration option is ignored.