cogd CFM run locking
When cogd performs a configuration management run, it should be the only
thing doing so.  If a one-off (cogd -1) run comes in while the scheduled
daemon is going, it will be ignored and a message printed to that
effect.  Likewise, if `cogd -1` is mid-run when the schedule says its
time, the scheduled run should get bumped.

This is done via a lock file, usually in /var/lock/cogd.

v2.0.9