Handle missing lockdir more robustly
Instead of claiming that another process is running, we should error if
we couldn't acquire the lock AND the lock file doesn't exist.  In
testing, this happened when lockdir didn't exist, but it could also
happen if the caller lacks permission to the lockdir, the lockdir isn't
a dir, etc.