Clockwork Mesh - Remote Orchestration Framework
Mesh is a remote execution and orchestration framework for Clockwork
that allows site administrators to run specific commands on all or a
subset of their Clockwork-managed hosts.

It supports a robust ACL implementation for allowing and denying access
to run sepcific commands on target hosts, based on configuration
management policy.  You can specify fragments of your ACL alongside
resources in the policy manifest, and they will be collected and
enforced as a group.

This is a fairly invasive change, and while it doesn't have a lot of
documentation (yet), and it doesn't do a whole lot (yet) it does provide
the foundation for some pretty powerful paradigms.