Handle both /etc/clockwork/cw.conf and ~/.cwrc
If no explicit configuration file is specified, treat ~/.cwrc as an
override to the global /etc/clockwork/cw.conf.  If either file doesn't
exist, treat that as an acceptable situation and continue on as if the
file in question specified no configuration.

If an explicit configuration file (-c) was specified, skip the magical
default entirely and just load the one file.  It is an error if this
file doesn't exist.