Internal Dependency Shortcuts
Dependencies, both forward (A depends on B) and reverse (B affects A)
can now be specified internal to the stanza declaration:

    package "base" { }
    service "serviced" {
        running: "yes"
        enabled: "yes"

        depends on package("base")
    }
    file "/etc/base.conf" {
        source: "/files/$path"

        depends on package("base")
        affects service("serviced")
    }

This cuts down on the heft of dependency-heavy manifests.

cw-shell picked up a new command, 'show order' that shows the order that
resources will be enforced, based on the internal dependency resolution
logic.  This was originally introduced for testing Clockwork itself, but
may prove useful for diagnostics and troubleshooting.