Better library includes (not just source pasting)
New `#include x` lexer grapheme will cause the lexer to halt parsing the
current stream, search the PENDULUM_INCLUDE paths for x.pn, and load
that.  Once that is done, those opcodes get spliced into the main list
of opcodes, with a special 0xfe marker for delineating object
boundaries.

pcov was updated to include or excludes blocks of the binary image based
on the 0xfe delineations, so that the coverage tests for stdlib can
include *just* stdlib, and not the tests themselves.