New `augeas.exists?` opcode
Like it's peer `augeas.find`, `augeas.exists?` looks for an entry in the
hierarchy and sets the accumulator to 0 if it is found.  Unlinke
`augeas.find`, however, `augeas.exists?` does not heap-allocate to
return the full found path.

Use of this in stdlib.pn is predicated on runtime 20150201 or greater.

v3.0.8