Offline/Cache Mode
cogd now supports cache/offline mode.  If a previous CFM run was
successfull, the Pendulum bytecode image that represents the policy will
be cached.  Subsequent runs that fail to connect to a clockd master can
then fallback on this cached policy for local enforcement of the "last
known good state"